Fix all -Wall warnings and enable -Werror
Mechanical (clean across Qt5/Qt6): same shape as showeq-daemon —
format casts to %lld+(long long) for qsizetype args, uint/int cast
for sign-compares in xmlconv/xmlpreferences/mapcore/guildshell,
QVariant::Type switch operand cast to int, dead-write removals,
drop unused ok_in_hex.

Real bugs (Quarm-specific in addition to the shared ones):
- everquest.h: MAX_BUFFS was 42 (Live constant) but charProfileStruct
  declares spellBuff buffs[15] (EQMac layout). The buff-iteration
  loops in player.cpp and messageshell.cpp ran past the array end —
  -Waggressive-loop-optimizations flagged "iteration 15 invokes UB".
  Set MAX_BUFFS=15 to match the actual array size; replay goldens
  regenerated since the OOB reads happened to emit phantom buff
  signals into the previous golden.
- messageshell.cpp::handleChannelMessage: strncpy of 63 bytes into a
  64-byte field tripped -Wstringop-truncation. Switched to memcpy
  since the dest is memset-zeroed and the trailing NUL is preserved.
- spellFaded / setuid / connect2 payload: same fixes as showeq-daemon.

Build: -Werror added; -Wno-error=stringop-overread for Qt5 qstring.h.

Verified: ctest 3/3, replay 2/2 PASS (goldens regenerated), Qt5 + Qt6
builds clean.

+++ src/daemonapp.cpp
@@ -693,8 +693,8 @@ void DaemonApp::loadZoneMap(const QString& shortZoneName)
         }
         import = true;
     }
-    qInfo("loaded map for zone '%s' (%d layer(s) from %s)",
-          qUtf8Printable(shortZoneName), files.size(),
+    qInfo("loaded map for zone '%s' (%lld layer(s) from %s)",
+          qUtf8Printable(shortZoneName), (long long)files.size(),
           qUtf8Printable(QFileInfo(files.first()).absolutePath()));
 }
